plot.plsone(plstools)	R Documentation

Plot Diagnostics for an plsone Object

Description

Nine plots (selectable by which.plot) are currently available: Observed vs Expected values, Normal QQ-plot of residuals, Representation of components th, Prediction vs the h^{th} component, Representation of DModXN, Representation of Weights, Representation of DModYN, Representationof Hotellin T2, Representation of correlations with the conponents.

Usage

## S3 method for class 'plsone':
plot(x, xax = 1, yax = 2, mfrow = NULL, which.plot =1:9,...)

Arguments

x 	 an object of class inheriting from 'plsone'

xax 	 the column number for the x-axis

yax 	 the column number for the y-axis

mfrow 	 a vector of the form 'c(nr,nc)', otherwise computed by as special own function n2mfrow

which.plot 	 a numeric vector containing the numbers of The selected plots (see details)

... 	 further arguments passed to or from other methods.

Details

The selected plots are drawn on a graphics device.
1: Observed vs Expected values
2: Normal QQ-plot of residuals
3: Representation of Components th (in 2 dimensions, by default xax=1 and yax=2).
4: Prediction vs the h^{th} component (by default xax=1).
5: Representation of DModXN
6: Representation of Weights (in 2 dimensions, by default xax=1 and yax=2).
7: representation of DModYN
8: Representationof Hotellin T2
9: Representation of correlations with the conponents (in 2 dimensions, by default xax=1 and yax=2)

Value

x is invisibly returned.

References

Tenenhaus M.(1998) La Regression PLS. Theorie et pratique. Technip, Paris.

See Also

plsone, qqnorm, s.arrow, s.label, plot.lm

Examples

require(pls)
data(yarn)
plstest <- plsone(density ~ NIR, nf=6, data = yarn,scale=FALSE)
plot(plstest)
plot(plstest,which.plot=1:2)
plot(plstest,which.plot=c(1:2,6))

[Package plstools version 1.0-5 Index]